+++ gradle.properties
@@ -46,7 +46,7 @@ usesMixins = true
 # Adds some debug arguments like verbose output and export
 usesMixinDebug = true
 # Specify the location of your implementation of IMixinPlugin. Leave it empty otherwise.
-mixinPlugin = 
+mixinPlugin =
 # Specify the package that contains all of your Mixins. You may only place Mixins in this package or the build will fail!
 mixinsPackage = mixins
 # Specify the core mod entry class if you use a core mod. This class must implement IFMLLoadingPlugin!
