--- src/main/java/com/mitchej123/hodgepodge/common/KeyBindingDuck.java
@@ -1,6 +1,0 @@
-package com.mitchej123.hodgepodge.common;
-
-public interface KeyBindingDuck {
-
-    void hodgepodge$updateKeyStates();
-}
