+++ src/main/java/com/mitchej123/hodgepodge/mixins/interfaces/KeyBindingExt.java
@@ -1,0 +1,6 @@
+package com.mitchej123.hodgepodge.mixins.interfaces;
+
+public interface KeyBindingExt {
+
+    void hodgepodge$updateKeyStates();
+}
