+++ src/main/java/com/mitchej123/hodgepodge/mixins/late/thaumcraft/MixinEntityGolemBase.java
@@ -13,7 +13,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.mitchej123.hodgepodge.common.ThaumcraftMixinMethods;
+import com.mitchej123.hodgepodge.mixins.hooks.ThaumcraftMixinMethods;
 
 import thaumcraft.common.entities.golems.EntityGolemBase;
 import thaumcraft.common.entities.golems.Marker;
