+++ src/main/resources/mixins.hodgepodge.json
@@ -1,7 +1,6 @@
 {
   "required": true,
   "minVersion": "0.8.3-GTNH",
-  "package": "com.mitchej123.hodgepodge.mixins",
   "refmap": "mixins.hodgepodge.refmap.json",
   "target": "@env(DEFAULT)",
   "compatibilityLevel": "JAVA_8"
